backend/requirements.txt
Django>=5.0,<6.0
# REST API
djangorestframework>=3.15,<4.0
# CORS for React frontend
django-cors-headers>=4.4,<5.0
# Rich-text editor
django-ckeditor>=6.7,<7.0
# Google OAuth
django-allauth>=0.61,<1.0
# Required by allauth Google provider
requests>=2.32,<3.0
# JWT for auth providers
PyJWT>=2.9,<3.0
# Crypto backend used by allauth/google + JWT handling
cryptography>=42.0,<43.0
# OpenAI API
openai>=1.40,<2.0
# Database driver
psycopg[binary]>=3.1,<4.0
# Production server
gunicorn>=21.2,<22.0
